Algo de Mejoras

+++ java/apuestas_backend/core/src/main/java/com/devonfw/application/apuestas_backend/tipomejoramanagement/logic/impl/usecase/UcManageTipoMejoraImpl.java
@@ -1,5 +1,6 @@
 package com.devonfw.application.apuestas_backend.tipomejoramanagement.logic.impl.usecase;
 
+import java.util.List;
 import java.util.Objects;
 
 import javax.inject.Named;
@@ -9,6 +10,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
 
+import com.devonfw.application.apuestas_backend.mejoramanagement.logic.api.to.MejoraCto;
+import com.devonfw.application.apuestas_backend.mejoramanagement.logic.api.to.MejoraSearchCriteriaTo;
+import com.devonfw.application.apuestas_backend.mejoramanagement.logic.api.usecase.UcFindMejora;
 import com.devonfw.application.apuestas_backend.tipomejoramanagement.dataaccess.api.TipoMejoraEntity;
 import com.devonfw.application.apuestas_backend.tipomejoramanagement.logic.api.to.TipoMejoraEto;
 import com.devonfw.application.apuestas_backend.tipomejoramanagement.logic.api.usecase.UcManageTipoMejora;
@@ -22,16 +26,34 @@ import com.devonfw.application.apuestas_backend.tipomejoramanagement.logic.base.
 @Transactional
 public class UcManageTipoMejoraImpl extends AbstractTipoMejoraUc implements UcManageTipoMejora {
 
+	private UcFindMejora ucFindMejora;
+
+	private MejoraSearchCriteriaTo mejoraSearchCriteriaTo;
+
 	/** Logger instance. */
 	private static final Logger LOG = LoggerFactory.getLogger(UcManageTipoMejoraImpl.class);
 
 	@Override
 	public boolean deleteTipoMejora(long tipoMejoraId) {
 
-		TipoMejoraEntity tipoMejora = getTipoMejoraRepository().find(tipoMejoraId);
-		getTipoMejoraRepository().delete(tipoMejora);
-		LOG.debug("The tipoMejora with id '{}' has been deleted.", tipoMejoraId);
-		return true;
+		List<MejoraCto> listMejoraCtos = dependenciesById(tipoMejoraId);
+		if(listMejoraCtos.isEmpty()) {
+			return false;
+		}
+		else {
+			TipoMejoraEntity tipoMejora = getTipoMejoraRepository().find(tipoMejoraId);
+			getTipoMejoraRepository().delete(tipoMejora);
+			LOG.debug("The tipoMejora with id '{}' has been deleted.", tipoMejoraId);
+			return true;
+		}
+	}
+
+	public List<MejoraCto> dependenciesById(long tipoMejoraId) {
+		return getUcFindMejora().findMejoraCtos(mejoraSearchCriteriaTo).getContent();
+	}
+
+	public UcFindMejora getUcFindMejora() {
+		return ucFindMejora;
 	}
 
 	@Override
